% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaytonClayton.Weibull.MLE.R
\name{ClaytonClayton.Weibull.MLE}
\alias{ClaytonClayton.Weibull.MLE}
\title{Parameter estimation based on the Clayton copula for serial dependence and the Clayton copula for dependent censoring with the Weibull distributions}
\usage{
ClaytonClayton.Weibull.MLE(subject, t.event, event, t.death, death, stageI, Weibull.plot,
                                  jackknife, plot, GOF, GOF.plot, rep.GOF, digit)
}
\arguments{
\item{subject}{a vector for numbers of subject}

\item{t.event}{a vector for event times}

\item{event}{a vector for event indicator (=1 if recurrent; =0 if censoring)}

\item{t.death}{a vector for death times}

\item{death}{a vector for death indicator (=1 if death; =0 if censoring)}

\item{stageI}{an option to select MLE or LSE method for the 1st-stage optimization}

\item{Weibull.plot}{if TRUE, show the Weibull probability plot}

\item{jackknife}{if TRUE, the jackknife method is used for estimate covariance matrix (default = TRUE)}

\item{plot}{if TRUE, the plots for marginal distributions are shown (default = FALSE)}

\item{GOF}{if TRUE, show the p-values for KS-test and CvM-test}

\item{GOF.plot}{if TRUE, show the model diagnostic plot}

\item{rep.GOF}{repetition number of parametric bootstrap}

\item{digit}{accurate to some decimal places}
}
\value{
A list with the following elements:
\item{Sample_size}{Sample size N}
\item{Case}{Count for event occurences}
\item{scale1}{Scale parameter for Weib(scale1, shape1)}
\item{shape1}{Shape parameter for Weib(scale1, shape1)}
\item{scale2}{Scale parameter for Weib(scale2, shape2)}
\item{shape2}{Shape parameter for Weib(scale2, shape2)}
\item{theta}{Copula parameter for the Clayton copula C_theta}
\item{alpha}{Copula parameter for the Clayton copula tilde(C)_alpha}
\item{COV}{Asymptotic covariance estimated by the jackknife method}
\item{KS}{Kolmogorov-Smirnov test statistics}
\item{p.KS}{P-values for Kolmogorov-Smirnov tests}
\item{CM}{Cramer-von Mises test statistics}
\item{p.CM}{P-values for Cramer-von Mises tests}
\item{Convergence}{Convergence results for each stage}
\item{Jackknife_error}{Count for error in jackknife repititions}
\item{Log_likelihood}{Log-likelihood values}
}
\description{
Perform two-stage estimation based on the Clayton copula C_theta for serial dependence and the Clayton copula tilde(C)_alpha for dependent censoring with the marginal distributions Weib(scale1, shape1) and Weib(scale2, shape2). The jackknife method estimates the asymptotic covariance matrix. Parametric bootstrap is applied while doing Kolmogorov-Smirnov tests and Cramer-von Mises test. The guide for using this function shall be explained by Huang (2019), and Huang, Wang and Emura (2020).
}
\details{
When jackknife=FALSE, the corresponding standard error and confidence interval values are shown as NA.
}
\examples{
data = ClaytonClayton.Weibull.data(N = 30, scale1 = 1, shape1 =0.5, theta = 2,
                                   scale2 = 0.45, shape2 = 0.5, alpha = 2, b = 10, l = 300)
\donttest{
ClaytonClayton.Weibull.MLE(subject = data$Subject,
                           t.event = data$T_ij, event = data$delta_ij,
                           t.death = data$T_i_star, death = data$delta_i_star,
                           jackknife= TRUE, plot = TRUE)
}

}
\references{
Huang XW, Wang W, Emura T (2020) A copula-based Markov chain model for serially dependent event times with a dependent terminal event. Japanese Journal of Statistics & Data Science. Accepted.
}
\author{
Xinwei Huang
}
