% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula.Markov.survival.R
\docType{package}
\name{Copula.Markov.survival}
\alias{Copula.Markov.survival}
\alias{Copula.Markov.survival-package}
\title{Copula.Markov.survival}
\description{
Perform likelihood estimation and corresponding analysis under the copula-based Markov chain model for serially dependent event times with a dependent terminal event. A two stage estimation method is applied for estimating model parameters. Two copula functions are used for measuring dependence. One is used for modeling serial dependence in recurrent event times. The other one is for modeling dependent censoring. The baseline hazard functions are modeled by the Weibull distributions. See Huang (2019) <https://etd.lib.nctu.edu.tw/cgi-bin/gs32/ncugsweb.cgi?o=dncucdr&s=id=%22GC106225601%22.&searchmode=basic> for detail.
}
\references{
Huang, X.-W. (2019). Likelihood-based inference for copula-based Markov chain models for continuous, discrete, and survival data. NCU library.
}
\author{
Xinwei Huang \email{xinweihuang@hotmail.com}
}
